moved mqtt util functions
