+++ dsh_sdk/src/utils.rs
@@ -177,29 +177,6 @@ pub(crate) fn get_env_var(var_name: &str) -> Result<String, DshError> {
     }
 }
 
-pub(crate) fn extract_header_and_payload(raw_token: &str) -> Result<&str, DshError> {
-    let parts: Vec<&str> = raw_token.split('.').collect();
-    parts
-        .get(1)
-        .copied()
-        .ok_or_else(|| DshError::ParseDnError("Header and payload are missing".to_string()))
-}
-
-pub(crate) fn decode_payload(payload: &str) -> Result<Vec<u8>, DshError> {
-    use base64::{alphabet, engine, read};
-    use std::io::Read;
-
-    let engine = engine::GeneralPurpose::new(&alphabet::STANDARD, engine::general_purpose::NO_PAD);
-    let mut decoder = read::DecoderReader::new(payload.as_bytes(), &engine);
-
-    let mut decoded_token = Vec::new();
-    decoder
-        .read_to_end(&mut decoded_token)
-        .map_err(DshError::IoError)?;
-
-    Ok(decoded_token)
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
